Bit of code hygiene, ensure casting to correct data type.

+++ modules/csg/csg_shape.cpp
@@ -378,7 +378,7 @@ static void _pack_manifold(
 		uint32_t reserved_id = r_manifold.ReserveIDs(1);
 		mesh.runOriginalID.push_back(reserved_id);
 		Ref<Material> material;
-		if (material_id < p_mesh_merge->materials.size()) {
+		if (material_id < (uint32_t)p_mesh_merge->materials.size()) {
 			material = p_mesh_merge->materials[material_id];
 		}
