+++ scene/resources/gradient.cpp
@@ -148,7 +148,7 @@ void Gradient::set_offsets(const Vector<float> &p_offsets) {
 }
 
 void Gradient::set_colors(const Vector<Color> &p_colors) {
-	if (points.size() < p_colors.size()) {
+	if (points.size() < (uint32_t)p_colors.size()) {
 		is_sorted = false;
 	}
 	points.resize(p_colors.size());
